Update hotel project

+++ HotelProject.Data/Repoistories/RoomsRepository.cs
@@ -1,45 +1,40 @@
 ﻿using HotelProject.Core.Repositories;
 using HotelProject.Core.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace HotelProject.Data.Repositories
 {
     public class RoomsRepository : IRoomsRepository
     {
         private readonly DataContext _context;
-
-        public RoomsRepository(DataContext context)
-        {
-            _context = context;
-        }
+        public RoomsRepository(DataContext context) => _context = context;
 
         public async Task<List<Room>> GetAllAsync()
         {
-            return await _context.Rooms.ToListAsync();
+            return await _context.Rooms
+                .Include(r => r.BookingRooms)
+                .ThenInclude(br => br.Booking) 
+                .ToListAsync();
         }
 
         public async Task<Room> GetByIdAsync(int id)
         {
-            return await _context.Rooms.FindAsync(id);
+            return await _context.Rooms
+                .Include(r => r.BookingRooms)
+                .ThenInclude(br => br.Booking)
+                .FirstOrDefaultAsync(r => r.Id == id);
         }
 
-        public void AddRoom(Room room)
-        {
-            _context.Rooms.Add(room);
-        }
+        public void AddRoom(Room room) => _context.Rooms.Add(room);
+
+        public async Task SaveAsync() => await _context.SaveChangesAsync();
 
         public async Task DeleteByIdAsync(int id)
         {
             var room = await GetByIdAsync(id);
-            if (room != null)
-            {
-                _context.Rooms.Remove(room);
-            }
+            if (room != null) _context.Rooms.Remove(room);
         }
 
         public async Task UpdateRoomAsync(int id, Room room)
@@ -47,12 +42,11 @@ namespace HotelProject.Data.Repositories
             var roomToUpdate = await _context.Rooms.FindAsync(id);
             if (roomToUpdate != null)
             {
-                roomToUpdate.Status = room.Status;
                 roomToUpdate.NumberOfBeds = room.NumberOfBeds;
                 roomToUpdate.BasePrice = room.BasePrice;
+                roomToUpdate.Status = room.Status;
             }
         }
-
         public async Task UpdateRoomStatusAsync(int id, RoomStatus roomStatus)
         {
             var roomToUpdate = await _context.Rooms.FindAsync(id);
@@ -60,11 +54,6 @@ namespace HotelProject.Data.Repositories
             {
                 roomToUpdate.Status = roomStatus;
             }
-        }
-
-        public async Task SaveAsync()
-        {
-            await _context.SaveChangesAsync();
         }
     }
 }
